refactor: renamed a module
