+++ packages/dantalion-i18n/src/resources/t.ts
